requirements.txt
# Core dependencies
langchain>=0.0.267
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
numpy>=1.24.3
python-dotenv>=1.0.0
pydantic>=2.0.0
backoff>=2.2.1
uuid>=1.30

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
mammoth>=1.6.0

# LLM and embedding
openai>=1.0.0
tiktoken>=0.4.0

# Web search
duckduckgo-search>=3.0.0

# UI
streamlit>=1.24.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0
